+++ feature/shared-preferences/src/commonMain/kotlin/williankl/accountkt/feature/sharedPreferences/models/CurrencyPreferences.kt
@@ -4,5 +4,5 @@ import williankl.accountkt.data.currencyService.models.Symbol
 
 public data class CurrencyPreferences(
     val preferredSymbol: Symbol,
-    val defaultAmount: Double,
+    val defaultAmount: Float,
 )
